+++ repos/os/src/server/event_filter/event_session.h
@@ -113,7 +113,7 @@ class Event_filter::Event_session : public Session_object<Event::Session, Event_
 
 		void submit_batch(unsigned const count)
 		{
-			size_t const max_events = _ds.size() / sizeof(Input::Event);
+			unsigned const max_events = (unsigned)(_ds.size() / sizeof(Input::Event));
 
 			if (count > max_events)
 				warning("number of events exceeds dataspace capacity");
